% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{aids_survival}
\alias{aids_survival}
\title{AIDS survival probability curve}
\usage{
aids_survival(n = 365 * 5)
}
\arguments{
\item{n}{Number of days to calculate}
}
\value{
A vector where the ith element is the probability that an
individual infected with HIV i days ago has not progressed to AIDS.
}
\description{
AIDS survival probability curve
}
\references{
Ron Brookmeyer and James J Goedert.  Censoring in an epidemic with an appli-cation to hemophilia-associated aids.Biometrics, pages 325–335, 1989.
}
