% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rita.R
\docType{data}
\name{assay_data}
\alias{assay_data}
\title{Example Assay Data}
\description{
Example Assay Data
}
\details{
1. `recent`: Logical. Did the subject test recent on the assay.
2. `hiv`: Logical. Is the subject HIV+.
3. `undiagnosed`: Logical. Has the subject never received a positive HIV test.
4. `elite_cntr`: Logical. Viral load test <1,000.
5. `tslt`: Numeric. Time since last HIV test in months.
6. `ever_hiv_test`: Logical. Has the subject never been tested for HIV.
7. `weights`: Numeric. Survey Weights.
8. `btwt001-btwt206`: Numeric. Jackknife replicate weights.
}
\keyword{data}
