% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen-survival.R
\name{diagnosis_survival}
\alias{diagnosis_survival}
\title{Survival probability for RITA Screening}
\usage{
diagnosis_survival(
  undiagnosed,
  tslt,
  ever_hiv_test,
  hiv,
  weights,
  n = 365 * 5,
  population = c("undiagnosed", "negative")
)
}
\arguments{
\item{undiagnosed}{A logical vector indicating if an individual has never been diagnosed as HIV+.}

\item{tslt}{Time since last test in days}

\item{ever_hiv_test}{A logical vector indicating whether an individual has ever had an HIV test}

\item{hiv}{HIV positive status.}

\item{weights}{Survey weights}

\item{n}{Number of days to calculate}

\item{population}{If undiagnosed, the testing histories of undiagnosed HIV+ people are used. If negative, the HIV- population is used.}
}
\value{
A vector where the ith element is the probability that an
individual infected i days ago has not been diagnosed.
}
\description{
Survival probability for RITA Screening
}
\examples{
  data("assay_data")
  diagnosis_survival(
    assay_data$undiagnosed,
    assay_data$tslt,
    assay_data$ever_hiv_test,
    assay_data$hiv,
    assay_data$weights,
    n=40)
}
