% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{lag_avidity_frr}
\alias{lag_avidity_frr}
\title{False Recency Rate (FRR) for the LAg-Avidity}
\usage{
lag_avidity_frr()
}
\value{
A vector of length 2. The first element is the proportion of untreated, non-elite controller, non-AIDS individuals infected more
than 2 years ago that test as recent on the LAg-Avidity assay. The second is the standard
error of the proportion.
}
\description{
False Recency Rate (FRR) for the LAg-Avidity
}
