% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{lag_avidity_survival}
\alias{lag_avidity_survival}
\title{Survival curve for testing recent}
\usage{
lag_avidity_survival(n = 365 * 5)
}
\arguments{
\item{n}{Number of days to calculate.}
}
\value{
A vector where the ith element is the probability that an untreated non-elite controller
individual infected i days ago tests recent on the LAg-Avidity assay.
}
\description{
Survival curve for testing recent
}
\details{
A reanalysis of Duong et. al. (2015). See Fellows (2021) for details
}
\references{
Yen T Duong, Reshma Kassanjee, Alex Welte, Meade Morgan, Anindya De, TrudyDobbs, Erin Rottinghaus, John Nkengasong, Marcel E Curlin, Chonticha Kitti-nunvorakoon,  et  al.   Recalibration  of  the  limiting  antigen  avidity  eia  to  deter-mine mean duration of recent infection in divergent hiv-1 subtypes.PloS  one,10(2):e0114947, 2015.
}
