% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{print.rita_boot_est}
\alias{print.rita_boot_est}
\title{print bootstrap}
\usage{
\method{print}{rita_boot_est}(x, ...)
}
\arguments{
\item{x}{a inc_boot_est object}

\item{...}{additional parameters for summary.inc_boot_est}
}
\description{
print bootstrap
}
