% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{rita_bootstrap}
\alias{rita_bootstrap}
\title{Survey bootstrap}
\usage{
rita_bootstrap(
  recent,
  undiagnosed,
  elite_cntr,
  hiv,
  tslt,
  ever_hiv_test,
  weights,
  rep_weights = NULL,
  rep_weight_type = c("BRR", "Fay", "JK1", "JK2", "JKn", "bootstrap", "other"),
  combined_weights = TRUE,
  diag_surv = NULL,
  tau = 2,
  frr = lag_avidity_frr()[1],
  assay_surv = lag_avidity_survival(tau * 365),
  aids_surv = aids_survival(tau * 365),
  show_progress = TRUE,
  ...
)
}
\description{
Survey bootstrap
}
