% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{summary.rita_boot_est}
\alias{summary.rita_boot_est}
\title{summary of incidence bootstrap}
\usage{
\method{summary}{rita_boot_est}(object, conf_level = 0.95, ...)
}
\arguments{
\item{object}{a inc_boot_est object}

\item{conf_level}{confidence level for intervals}

\item{...}{additional parameters for print.data.frame}
}
\description{
summary of incidence bootstrap
}
