% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen-survival.R
\name{diagnosis_survival}
\alias{diagnosis_survival}
\title{Survival probability for RITA Screening}
\usage{
diagnosis_survival(undiagnosed, tslt, ever_hiv_test, weights, n = 365 * 5)
}
\arguments{
\item{undiagnosed}{A logical vector indicating if an individual has never been diagnosed as HIV+.}

\item{tslt}{Time since last test in days}

\item{ever_hiv_test}{A logical vector indicating whether an individual has ever had an HIV test}

\item{weights}{Survey weights}

\item{n}{Number of days to calculate}
}
\value{
A vector where the ith element is the probability that an
individual infected i days ago has been diagnosed.
}
\description{
Survival probability for RITA Screening
}
