% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{rita_incidence}
\alias{rita_incidence}
\title{Assay Based Incidence Estimation}
\usage{
rita_incidence(
  recent,
  undiagnosed,
  elite_cntr,
  hiv,
  weights,
  tslt = NULL,
  ever_hiv_test = NULL,
  diag_surv = NULL,
  tau = 2 * 365,
  frr = lag_avidity_frr()[1],
  frr_se = lag_avidity_frr()[2],
  assay_surv = lag_avidity_survival(),
  aids_surv = aids_survival(tau)
)
}
\arguments{
\item{recent}{Logical. Tests recent on assay}

\item{undiagnosed}{Logical. No previous diagnosis}

\item{elite_cntr}{Logical. Is an elite controller (VL < 1000)}

\item{hiv}{Logical. Is HIV positive}

\item{weights}{Survey weights}

\item{tslt}{Time since last HIV test}

\item{ever_hiv_test}{Subject has been tested for HIV in the past}

\item{diag_surv}{time to diagnosis survival function vector}

\item{tau}{long term cut-off}

\item{frr}{False recency rate among treatment naive non-elite controller non-AIDS individuals}

\item{frr_se}{Standard error of FRR}

\item{assay_surv}{Survival function vector for assay among treatment naive non-elite controller non-AIDS individuals}

\item{aids_surv}{Survival function vector for time to AIDS in untreated individuals.}
}
\description{
Assay Based Incidence Estimation
}
